+++ libs/react/fpkit/src/App.tsx
@@ -1,7 +1,8 @@
 import React, { useState } from 'react'
 import './App.css'
 
-import FPBadge from './component/badge/fp-badge'
+import FPBadge from './components/badge/fp-badge'
+import { Box } from './index'
 
 function App() {
   const [count, setCount] = useState(0)
@@ -13,6 +14,9 @@ function App() {
         <p>
           <FPBadge elm="span" role='note'>Badge</FPBadge>
         </p>
+        <Box as='section'>
+          <FPBadge elm="span" role='note'>Badge</FPBadge>
+        </Box>
       </header>
     </div>
   )