+++ libs/react/fpkit/src/types/shared.ts
@@ -1,0 +1,6 @@
+export interface ComponentProps {
+  children?: React.ReactNode
+  renderStyles?: boolean
+  styles?: {}
+  classes?: String
+}