+++ libs/react/fpkit/CHANGELOG.md
@@ -3,49 +3,102 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
-# [0.1.0](https://github.com/shawn-sandy/react-vite/compare/@fpkit/fp@0.0.5...@fpkit/fp@0.1.0) (2022-09-29)
-
+## [0.2.11](https://github.com/shawn-sandy/fpkit/compare/@fpkit/fp@0.2.10...@fpkit/fp@0.2.11) (2022-11-18)
 
 ### Features
 
-* ✨ make span conditional add default span tag ([c4c961a](https://github.com/shawn-sandy/react-vite/commit/c4c961a3e402a6343771a0415ac04328227fc6ac))
+- 👆 update disabled button styles ([efad4b2](https://github.com/shawn-sandy/fpkit/commit/efad4b2a61010142974dc5178a079e394e33cec3))
 
+## [0.2.10](https://github.com/shawn-sandy/fpkit/compare/@fpkit/fp@0.2.9...@fpkit/fp@0.2.10) (2022-11-18)
 
+**Note:** Version bump only for package @fpkit/fp
 
+## [0.2.9](https://github.com/shawn-sandy/fpkit/compare/@fpkit/fp@0.2.8...@fpkit/fp@0.2.9) (2022-11-05)
 
+**Note:** Version bump only for package @fpkit/fp
+
+## [0.2.8](https://github.com/shawn-sandy/fpkit/compare/@fpkit/fp@0.2.7...@fpkit/fp@0.2.8) (2022-10-20)
+
+### Features
+
+- 🚧 update dropdown component ([80d5006](https://github.com/shawn-sandy/fpkit/commit/80d5006607064cd80b35253ba70f706aa9bc5c91))
+
+## [0.2.7](https://github.com/shawn-sandy/fpkit/compare/@fpkit/fp@0.2.6...@fpkit/fp@0.2.7) (2022-10-14)
+
+### Features
+
+- 🚧 update dropdown component ([b349385](https://github.com/shawn-sandy/fpkit/commit/b349385ec82bc08f5c164720074e69cc2323fa43))
+
+## [0.2.6](https://github.com/shawn-sandy/fpkit/compare/@fpkit/fp@0.2.5...@fpkit/fp@0.2.6) (2022-10-14)
+
+### Features
+
+- 🚧 update card props values ([e852d27](https://github.com/shawn-sandy/fpkit/commit/e852d2715a8597c132ccbe49c7b17615cf51d8db))
+
+## [0.2.5](https://github.com/shawn-sandy/fpkit/compare/@fpkit/fp@0.2.4...@fpkit/fp@0.2.5) (2022-10-12)
+
+**Note:** Version bump only for package @fpkit/fp
+
+## [0.2.4](https://github.com/shawn-sandy/fpkit/compare/@fpkit/fp@0.2.3...@fpkit/fp@0.2.4) (2022-10-12)
+
+### Features
+
+- add a box component ([b445b5a](https://github.com/shawn-sandy/fpkit/commit/b445b5a5920b5fc892a758089b5ad325cec5bb40))
+
+## [0.2.3](https://github.com/shawn-sandy/fpkit/compare/@fpkit/fp@0.2.2...@fpkit/fp@0.2.3) (2022-10-09)
+
+### Features
+
+- 📝 storybooks updates ([86b3662](https://github.com/shawn-sandy/fpkit/commit/86b3662224e7f259ba4452622d6566c42a647963))
+
+## [0.2.2](https://github.com/shawn-sandy/fpkit/compare/@fpkit/fp@0.2.1...@fpkit/fp@0.2.2) (2022-10-02)
+
+**Note:** Version bump only for package @fpkit/fp
+
+## [0.2.1](https://github.com/shawn-sandy/fpkit/compare/@fpkit/fp@0.2.0...@fpkit/fp@0.2.1) (2022-10-02)
+
+### Features
+
+- 🚧 base dropdown component ([902d945](https://github.com/shawn-sandy/fpkit/commit/902d945fd9100d5bdc266b36edad248b1a7e542f))
+- 🚧 base dropdown component ([aff9e81](https://github.com/shawn-sandy/fpkit/commit/aff9e810866f16ff4d716032c8d30f0703671abc))
+- 🚧 scaffold dropdown-details component ([3233a88](https://github.com/shawn-sandy/fpkit/commit/3233a8832085b8bcbee1d6fc97b1f1f3294b2a32))
+
+# [0.2.0](https://github.com/shawn-sandy/fpkit/compare/@fpkit/fp@0.1.1...@fpkit/fp@0.2.0) (2022-09-30)
+
+### Features
+
+- ❇️ add a card component to fpkit ([e6238f8](https://github.com/shawn-sandy/fpkit/commit/e6238f860ab2458f83c00971d3e098c2e541ec58))
+
+## [0.1.1](https://github.com/shawn-sandy/fpkit/compare/@fpkit/fp@0.1.0...@fpkit/fp@0.1.1) (2022-09-30)
+
+### Features
+
+- 🏗️ add button component ([8bfdf87](https://github.com/shawn-sandy/fpkit/commit/8bfdf87ce5f8e84bd4e3c51efebcf6a8664f4e16))
+
+# [0.1.0](https://github.com/shawn-sandy/react-vite/compare/@fpkit/fp@0.0.5...@fpkit/fp@0.1.0) (2022-09-29)
+
+### Features
+
+- ✨ make span conditional add default span tag ([c4c961a](https://github.com/shawn-sandy/react-vite/commit/c4c961a3e402a6343771a0415ac04328227fc6ac))
 
 ## [0.0.5](https://github.com/shawn-sandy/react-vite/compare/@fpkit/fp@0.0.4...@fpkit/fp@0.0.5) (2022-09-29)
 
-
 ### Features
 
-* ✨ set renderstyles to stur by default ([9a466ce](https://github.com/shawn-sandy/react-vite/commit/9a466ce583f87e9bccdd8ba11203b961542faa4a))
-
-
-
-
+- ✨ set renderstyles to true by default ([9a466ce](https://github.com/shawn-sandy/react-vite/commit/9a466ce583f87e9bccdd8ba11203b961542faa4a))
 
 ## [0.0.4](https://github.com/shawn-sandy/react-vite/compare/@fpkit/fp@0.0.3...@fpkit/fp@0.0.4) (2022-09-29)
 
 **Note:** Version bump only for package @fpkit/fp
 
-
-
-
-
 ## [0.0.3](https://github.com/shawn-sandy/react-vite/compare/@fpkit/fp@0.0.2...@fpkit/fp@0.0.3) (2022-09-29)
 
 **Note:** Version bump only for package @fpkit/fp
 
-
-
-
-
 ## 0.0.2 (2022-09-29)
-
 
 ### Features
 
-* 🚧 ad badge component fpkit/fp ([c8b6b7b](https://github.com/shawn-sandy/react-vite/commit/c8b6b7b11951ce8586505a01b2658b79187e3828))
-* 🚧 ad badge component pf fpkit/fp ([f17fd2d](https://github.com/shawn-sandy/react-vite/commit/f17fd2d61cf65cd6150e7cda77debe880cca266a))
-* 🚧 add new component ([fcb8001](https://github.com/shawn-sandy/react-vite/commit/fcb800111d2a92b063c26e8049c51c4d2e503887))
+- 🚧 ad badge component fpkit/fp ([c8b6b7b](https://github.com/shawn-sandy/react-vite/commit/c8b6b7b11951ce8586505a01b2658b79187e3828))
+- 🚧 ad badge component pf fpkit/fp ([f17fd2d](https://github.com/shawn-sandy/react-vite/commit/f17fd2d61cf65cd6150e7cda77debe880cca266a))
+- 🚧 add new component ([fcb8001](https://github.com/shawn-sandy/react-vite/commit/fcb800111d2a92b063c26e8049c51c4d2e503887))